File offsets should be off_t, not int32, for large file support.  Per OpenSuSE patch:  https://build.opensuse.org/package/view_file/security/tripwire/tripwire-2.4.1.2-src.off_t.patch?expand-1

+++ src/core/file.h
@@ -73,7 +73,7 @@ class  cFile
 {
 public:
 #if IS_UNIX
-	typedef int32 File_t;
+	typedef off_t File_t;
 
 #else //WIN32
     typedef int64 File_t;
